requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
tldextract==5.1.1
colorama==0.4.6
selenium
pillow==9.3.0
weasyprint==60.2
dnspython==2.4.2
Pillow==9.3.0
